Create request.py
Create request.py abstract class module.

Refactoring and moving all of the work from interaction.py into
request.py and subsuquent files completionrequest.py and editrequest.py.
The Request class is a base class meant to be used to create
CompletionRequest and EditRequest classes.

This makes working with the requests a bit easier and some of the code can be
reused.