Update window.py purpose
Remove functions from window.py and replace with sample function to
create a start window from a StartFrame class. Add self.frame member
variable and redefine class purpose in class comment.

This fundamentally changes how the Window class will operate within the
scope of the project. Now, a window class instantiates a window object
that holds a single frame. This frame is from a custom Frame class that
builds on other frames and widgets to create the window the user
interacts with. If the window needs to change how it looks, it just
creates a new frame and then packs the frame into the window.

Instead of creating multiple derived classes of the Window class and
managing multiple windows, instead there will be a frame base class with
children frame classes adding components to the frame. This allows for
easy reusability and extensibility within the program.