Create settingsframe.py
Create SettingsFrame class, derived class of BaseFrame. Add functions
for class.

This class will work as the parent class to the different lists of
settings that need to be applied when doing a completion or edit request
to OpenAI ChatGPT models, though it can work with any dictionary of
settings.