Refactor getSettings() to CompletionRequest
Refactor getCompletionSettings() to CompletionRequest.

This puts the responsibility of getting the request settings back into
the court of CompletionRequest.

For future: getSettings() can be refactored into Request and the bodies
of getEditSettings() and getCompletionSettings() can be refactored into
a single getSettings(self, class) function in the Controller class.
Should simplify the codebase a bit.