Replace empty kwargs in SettingsFrame
Replace the empy kwargs initialization with the creation of an output
variable, the setting of the default value, putting the output in an
outputs dictionary, and setting kwargs["variable"] equal to the output.

This gives the program a way to read the output from whatever widget the
settings window is using. This could be refactored into another
function.