Remove [option] from get() and set() in Request
Remove the [option] from the get() and set() functions in the Request
class.

This was used when the dictionary was split into optional and required,
but it no longer is. This makes using [option] throw an error. This also
means that the getters and setters for the specific classes can be
removed, as this is a more usable set of methods.