Remove getters and setters from CompletionRequest
Removed all getters and setters except for getResponse() and
addToPrompt().

These functions are obsolete since Request has get() and set().
getResponse() is required for the class to work and addToPrompt() allows
the user to add a message to a previous prompt, which is differet from
the regular get() and set(), though it could be replicated with them.