Create test_models.py
Create test_models.py and write unit tests for the class.

This tests the Models class and makes sure the object was set up
correctly (and that the functions work). Something that should be done
is split up the tests. Some of the tests contain too many assertions and
need to be done separately from one another.