Reorder assertIn() and assertNotIn() arguments
Reorder assertIn() and assertNotIn() arguments for the
test_ModelsFilterCorrectly() function in test_models.py.

The tests need to check that the string is in/not in the model name; the
tests were checking the other way around, that the model name was in/not
in the string we were filtering out. This corrects that.