Create empty TestController class
I'm not sure how I want to test this class yet, much of the code doesn't
work because I have been jumping around trying to write everything. The
goal is to get the rest of the code working then figure out what needs
to be in the Controller class and write tests for that.