Rewrite SettingsInputFrame class
Rewrite the class to create the frame upon initialization of the object,
to accept more inputs for kwargs, and to add the widgets to the frame.

This greatly simplifies the use of the SettingsInputFrame class. The
frame is entirely managed by the class and the object can then be
attached to its parent frame by the parent frame. Just pass the
functions and variables and then attach the frame to the settings
window.