Rewrite SettingsFrame class for new SettingsInputFrame
Rewrite the class to accept the new SettingsInputFrame objects.

The match structure was changed along with the functions that the match
structure uses. The settings functions are updated to better use the
SettingsInputFrame objects.

This interaction between the frames will be the defacto method of
interaction for all frames going forward. This simplifies the process, I
think.