Create destroy() method in BaseFrame class
Create destroy() method in BaseFrame that calls self.master.destroy().

This allows any frame based on the BaseFrame class to call the destroy()
function to close the window. As frames are built up on top of each
other they will all point back to different frames until they reach the
common frame that is attached to the window. This will continue to call
the destroy() until the window is reached and its destroy() is
called and the window is closed.