Add save and exit methods to SettingsFrame
Add saveSettings() and destroy() methods to SettingsFrame save and exit
functions.

These will let the settings window save the settings that it has created
and also close the window to not save the settings. The only issue is
that the saveSettings() method should be defined within the child
settings classes, but for now I'll leave it and come back to fix it
later.