Refactor settingsinputframe.py
Remove main from super().__init__(). Remove self.main.

Main is handled by tk.Frame, so no sense in having it here. Also, had
too many arguments in the super().__init__() call so I removed the
reference to main.

Remove label from addSettingsWidget() and change reference to label to
self.label.

This simplifies the call to addSettingsWidget(). We are already passing
label to __init__(), so just use the object member variable instead.

Remove label and defatul from createOutput() and relable default, label
as self.default, self.label.

This is for the same reason as the refactoring to addSettingsWidget().
This just simplifies the calls to the function.