Move TestResponse derived class into TestResponse
Move the TestResponse(unittest.TestCase) class in test_response.py into
the class TestResponse; create TestResponse class.

By moving the class inside a class that is not derived from
unittest.TestCase, the class will not run unless inherrited. This fixes
the errors that the class was throwing since it is not designed to run
on its own but to be a base class for more specific reponse test cases.