Rename classes and fix inheritance for test_response.py
Rename TestResponse (parent class) to TestResponseWrapper. Fix
inheritance for TestResponseTextCompletion and TestResponseEdit to
inherit from TestReponseWrapper.TestResponse.

This fixes the issue of the classes not running any tests and makes it
easier to understand what classes are doing what; however, now I need to
fix where tests are getting their data because they are not using the
variables correctly.