Remove labelWdiget and widget attachments from baseSetting()
This removes the grid attachments from the labelWdiget and widget
variables in the baseSetting() method in the settingsFrame class. This
should be directly handled by the SettingsInput class and not the
SettingsFrame class.

There is too much inter-class play here and it needs to be tightened up
a bit; it's difficult to follow.