Change master from frame to self in baseSetting()
Change labelWidget and widget master arguments from the frame object to
self.

This moves the widgets from the independent settingsinputframe object to
the settingsframe object. Necessary to fix the grid placements issues
with the settings.

The widgets are not properly showing up, but that is only because they
are being overwritten. This will be fixed soon.