Add exception handling for create() in settingsframe.py
Add exception handling for settings of unaccounted for types. Throws
error, prints error, and continues through the settings.

This should handle any time a setting is used that is not related to
one of the functions that can handle the input type. It will just skip
over the setting and continue through the list, printing the output of
the error to the command prompt.