Refactor match structure to getSettings()
Refactor the match decision structure from create() to getSettings() in
settingsframe.py.

The method getSettings() now returns the function to be used in creating
the widgets. This function is then used in a more general way to create
the labelWidget and widget objects, which are then attached to the
current frame.

Everything still works.