Add body to saveSettings() method in settingsframe.py
This body loops through the settings used to setup the frame and stores
the values being inputted into the settings dictionary. If it finds a
value that is not in the outputs but in the dictionary it will pass over
it.

At the moment it is causing an issue, but that is because I have not
written out the initialization of the tk variables. Will do that next.