Fix response assignment in sendInput()
Fix controller.response to access private variable since there is no
setter function for the _response variable. Then, add test from the
response to the completion input prompt.

This works, but the method for adding text to the prompt does not work
how I would like it to work. Instead of insert it should be a replace
method.