Add binding to _prompt for underlining
Add a binding to the Text input for underlining text as a user types.

This is just a test of the system to figure out how it works. The
binding has to be done to the Text object and the tag group has to be
updated constantly to get it to work. That it works is good.