Update self._requestDict to use model object for "model"
Use the model object instead of the name of the model in the "model"
value for the request dictionary in CompletionRequest. Change set_model()
to use model objects.

This should make it easier to access information about the models. This
breaks a lot of tests, so they will all need to be rewritten.