Fix errors in settingsframe.py
Fix errors related to saving the settings. This required reworking
functions in completion.py and removing tests from
test_completionrequest.py.

This allows settings to be saved; however, it is reliant on the settings
being request settings. This is not desired and I would like to abstract
this out some more. Still a bit of work to do here to get it how I want
it, but it is working.