Refactor from saveSettings() to setAttr()
Refactor settings-specific setting code from saveSettings() to setAttr()
in completionsettings.py.

This moves the statemenets selecting and setting different attributes to
the class that inherits from SettingsFrame class. setAttr() returns a
function defined in the body of setAttr() that handles setting the
values.

An improvement could be moving the function definition out of the body
of setAttr() and just using it to return the function.

Everything works.