Refactor setAttr() to not return function
Refactor setAttr() to not return a function but be the setter function.

This really simplifies the process even more. I don't know why I thought
it returning the setter function was a good idea. Just have it be the
setter function and that simplifies the process a lot.

Everything works.