Refactor getSettings() to use getattr()
Refactor in settingsframe.py and rename modelSetting() to ModelSetting.

This greatly simplifies getting the different setting functions for
create(). In fact, this entire function could almost be refactored and
not even used. Something to consider. The advantage of having it is
having the exception raising hidden from the create() method, making
that method easier to read.