Refactor kwargs setup to _kwargs in settingsframe.py
This now sets up the variable key for kwargs, which is needed for every
input we have. Now using self.outputs[key] to reference the tkVar used
by the input outside of the _kwargs function.

This change only applies to ModelSetting() method. Working so far, so
should be able to rewrite to all methods soon.