Fix _kwargs error in settingsframe.py
Remove for-loop over inspect parameters and replace with new varKey
parameter in _kwargs() call.

Moving the varKey to an input parameter for the function puts the
responsbility of using the correct varKey on the function setting up the
setting. This removes the complexity of having to figure it out within
_kwargs().

Considering moving the set() call out of _kwargs to allow the setting
methods to handle that as well.