Fix _kwargs() set() call
Keep the set() call for kwargs[key] and instead require a redefinition of
the default value outside of _kwargs and in the setting function if
extra steps are needed for the display of the default value.

This fixes the issue I was having, though it could look better. This
project is due for a massive refactor push soon.