Move ModelSetting to completionsettings.py
This setting function is directly related to completion and edit
settings, so I moved it to the completionsetting.py module. Eventually
there will be a requestsetting.py that will hold shared methods between
the two settings and this method will move there, but for now just keep
it as-is until I do that.