Refactor creating menu in MainMenu class
Add createMenu() method, which creates the menu. Add addFileMenu()
method, which attaches the file menu to the menu. Change
createFileMenu() to output actual menu object and not MainFileMenu
object.

This should help simplify the process of generating menus? I'm not sure
yet and will continue to play around with it.