Add start() to app.py and remove openai import
Add controller.start() to the app.py executable and remove the openai
import temporarily.

The openai import doesn't work for me at the moment (because I'm not on
my home computer), so it'll have to wait until I get home to fully test.
So far, though, it is going in the right direction.