Refactor countLines to be more accurate
Change countLines to now use text.count() for counting the lines. Update
the heights of all text boxes to use this new method. Move text height
out of createText() method into create() method.

Now this allows for all text boxes' heights to be dynamically adjusted
to match the text in the box.