Make adjustments to API Frame methods
Add method for setting the API key and make some adjustments to other
methods.

This method for saving the API key does not work. I will likely have to
go back to storing the key in a file to be read by the program. I was
trying to export it to the environment variables but it just doesn't
work.