Change request to copy _requestDict in getResponse()
The fix I implemented was not coying the dictionary, so making a
request rewrote the model in the dictionary and broke the code. The
change fixes that and it is working properly.

The only thing left to fix is how the software interacts with the API
key. Right now it can't set the API variable so I'll have to create
another method.