Create option to load defaults from file in request.py
I still have the defaults written in there, but it now should load the
data from a pickle file. The goal is to have the pickle files be
generated on startup the first time the program is run and then just
used from there on out (no downloading suspicious pickle files).

This just feels like it would clean up the code a bit? To have it load
from a file instead of having it all written out in the class. This
could be done better.