Update apiWindow() to not create new frame if frame exists
Update apiWindow() method to check if self.frame exists. If it does,
then it does not create a new frame and just creates the window.
Otherwise, it creates the main frame and then generates the window.

This should let the user change the API key without having to restart
the program.